+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 2.3.4
 - Allow disabling of cache preloaded localized pricing
+- Allow disabling of Magento discount support in Flow Checkout
+- Address bug in daily sync configuration
 
 ## 2.3.3
 - Import latest label_id, commerical_invoice, pdf, zpl, and center to orders on label_upserted event